+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ## 0.0.41
-- [scoreboard] Added Numeric display & team
+- [scoreboard] Add numeric & team display
+- [state] Add LazyWatcher
 
 ## 0.0.40
 - [spigot] Add NMSUtils